update "manage data crossref"
